pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name =  "skoufas-library"
authors = [
    {name = "Claudio Bantaloukas", email = "rockreamer@gmail.com"},
]
description = "Django Web Application for managing Skoufas library catalog"
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Other Audience",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
]
requires-python = ">=3.10"
dynamic = ["version"]
dependencies = [
    "django-bootstrap-v5",
    "django-extensions",
    "django",
    "djangoql",
    "psycopg2",
    "python-stdnum",
    "pytz",
    "pyyaml",
]

[project.optional-dependencies]
test = [
    "autopep8==2.0.1",
    "bandit[toml]==1.7.4",
    "black==23.1.0",
    "check-manifest==0.49",
    "django-stubs[compatible-mypy]==1.14.0",
    "djlint==1.19.14",
    "flake8-bugbear==23.2.13",
    "flake8-docstrings",
    "flake8-formatter_junit_xml",
    "flake8==6.0.0",
    "mypy==0.991",
    "pre-commit==3.0.4",
    "pylint_junit",
    "pylint==2.16.2",
    "pytest-cov==4.0.0",
    "pytest-github-actions-annotate-failures",
    "pytest-mock<3.10.1",
    "pytest-runner",
    "pytest==7.2.1",
    "reorder-python-imports",
    "shellcheck-py==0.9.0.2",
    "toml",
]

[tool.hatch.version]
path = "skoufas_library_project/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
  "/",
]
exclude = [
    "/.vscode",
    "/junk",
    "/scripts",
    "/venv",
]

[tool.django-stubs]
django_settings_module = "skoufas_library_project.settings"

[tool.mypy]
plugins = [
    "mypy_django_plugin.main",
]
exclude =  '''(?x)(
    migrations    # files named "one.py"
    | two\.pyi$  # or files ending with "two.pyi"
    | ^three\.   # or files starting with "three."
)'''

[project.scripts]
skoufas-library = "manage:main"

[project.urls]
Documentation = "https://github.com/skoufas/skoufas-library#readme"
Source = "https://github.com/skoufas/skoufas-library"
Tracker = "https://github.com/skoufas/skoufas-library/issues"

[tool.black]
line-length = 120
fast = true

[tool.bandit]
exclude_dirs = ["build","dist", "junk", "tests","scripts"]
number = 4
recursive = true
targets = [
    "books",
    "loaning",
    "skoufas_library_project",
    "manage.py",
]

reportMissingImports = true
reportMissingTypeStubs = false

pythonVersion = "3.11"
pythonPlatform = "Linux"

# [tool.pytest.ini_options]
# #addopts = "--cov-report xml:coverage.xml --cov src --cov-fail-under 0 --cov-append -m 'not integration'"
# pythonpath = [
#   "src"
# ]
# testpaths = "tests"
# junit_family = "xunit2"
# markers = [
#     "slow: marks tests as slow",
#     "integration: marks tests as integration",
#     "unit: fast offline tests",
# ]

# [tool.tox]
# legacy_tox_ini = """
# [tox]
# envlist = py

# [testenv]
# commands =
#     pytest {posargs}
# """

[tool.pylint]
extension-pkg-whitelist= [
]
ignore="CVS,migrations"
ignore-patterns="test.*?py,conftest.py"
init-hook='import sys; sys.setrecursionlimit(8 * sys.getrecursionlimit())'
jobs=0
limit-inference-results=100
persistent="yes"
suggestion-mode="yes"
unsafe-load-any-extension="no"

[tool.pylint.'MESSAGES CONTROL']
enable="c-extension-no-member"
disable=[
    "too-few-public-methods",
    "no-else-return",
    "import-outside-toplevel",
    "no-member",
]

[tool.pylint.'REPORTS']
evaluation="10.0 - ((float(5 * error + warning + refactor + convention) / statement) * 10)"
output-format="text"
reports="no"
score="yes"

[tool.pylint.'REFACTORING']
max-nested-blocks=5
never-returning-functions="sys.exit"

[tool.pylint.'BASIC']
argument-naming-style="snake_case"
attr-naming-style="snake_case"
bad-names= [
    "foo",
    "bar"
]
class-attribute-naming-style="any"
class-naming-style="PascalCase"
const-naming-style="UPPER_CASE"
docstring-min-length=-1
function-naming-style="snake_case"
good-names= [
    "i",
    "j",
    "k",
    "ex",
    "Run",
    "_"
]
include-naming-hint="yes"
inlinevar-naming-style="any"
method-naming-style="snake_case"
module-naming-style="any"
no-docstring-rgx="^_"
property-classes="abc.abstractproperty"
variable-naming-style="snake_case"

[tool.pylint.'FORMAT']
ignore-long-lines="^\\s*(# )?.*['\"]?<?https?://\\S+>?"
indent-after-paren=4
indent-string='    '
max-line-length=120
max-module-lines=1000
single-line-class-stmt="no"
single-line-if-stmt="no"

[tool.pylint.'LOGGING']
logging-format-style="old"
logging-modules="logging"

[tool.pylint.'MISCELLANEOUS']
notes= [
    "FIXME",
    "XXX",
    "TODO"
]

[tool.pylint.'SIMILARITIES']
ignore-comments="yes"
ignore-docstrings="yes"
ignore-imports="yes"
min-similarity-lines=7

[tool.pylint.'SPELLING']
max-spelling-suggestions=4
spelling-store-unknown-words="no"

[tool.pylint.'STRING']
check-str-concat-over-line-jumps="no"

[tool.pylint.'TYPECHECK']
contextmanager-decorators="contextlib.contextmanager"
generated-members="numpy.*,np.*,pyspark.sql.functions,collect_list"
ignore-mixin-members="yes"
ignore-none="yes"
ignore-on-opaque-inference="yes"
ignored-classes="optparse.Values,thread._local,_thread._local,numpy,torch,swagger_client"
ignored-modules="numpy,torch,swagger_client,netCDF4,scipy"
missing-member-hint="yes"
missing-member-hint-distance=1
missing-member-max-choices=1

[tool.pylint.'VARIABLES']
additional-builtins="dbutils"
allow-global-unused-variables="yes"
callbacks= [
    "cb_",
    "_cb"
]
dummy-variables-rgx="_+$|(_[a-zA-Z0-9_]*[a-zA-Z0-9]+?$)|dummy|^ignored_|^unused_"
ignored-argument-names="_.*|^ignored_|^unused_"
init-import="no"
redefining-builtins-modules="six.moves,past.builtins,future.builtins,builtins,io"

[tool.pylint.'CLASSES']
defining-attr-methods= [
    "__init__",
    "__new__",
    "setUp",
    "__post_init__"
]
exclude-protected= [
    "_asdict",
    "_fields",
    "_replace",
    "_source",
    "_make"
]
valid-classmethod-first-arg="cls"
valid-metaclass-classmethod-first-arg="cls"

[tool.pylint.'DESIGN']
max-args=5
max-attributes=7
max-bool-expr=5
max-branches=12
max-locals=15
max-parents=7
max-public-methods=20
max-returns=6
max-statements=50
min-public-methods=2

[tool.pylint.'IMPORTS']
allow-wildcard-with-all="no"
analyse-fallback-blocks="no"
deprecated-modules="optparse,tkinter.tix"

[tool.pylint.'EXCEPTIONS']
overgeneral-exceptions= [
    "builtins.BaseException",
    "builtin.Exception"
]
